backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.30.1
PyMuPDF==1.23.14
requests>=2.28.0